+++ lib/preface_specs/preface_specs.mli
@@ -17,6 +17,10 @@
 module Semigroup = Semigroup
 module Monoid = Monoid
 
+(** {1 Lattice hierarchy} *)
+
+module Meet_semilattice = Meet_semilattice
+
 (** {1 Functor hierarchy} *)
 
 module Invariant = Invariant
@@ -91,6 +95,7 @@ module Freer_monad = Freer_monad
 
 module type SEMIGROUP = Semigroup.API
 module type MONOID = Monoid.API
+module type MEET_SEMILATTICE = Meet_semilattice.API
 module type FUNCTOR = Functor.API
 module type BIFUNCTOR = Bifunctor.API
 module type PROFUNCTOR = Profunctor.API